requirements.txt
asgiref==3.8.1
Django==5.1.7
django-environ==0.11.2
psycopg2-binary==2.9.7
PyJWT==2.8.0
sqlparse==0.5.3
tzdata==2025.1